Check if dir exists, reset dirs list.

+++ src/toil/cwl/cwltoil.py
@@ -990,7 +990,9 @@ def cleanTempDirs(job):
     """Remove temporarly created directories."""
     if job._succeeded:
         for tempDir in job.openTempDirs:
-            shutil.rmtree(tempDir)
+            if os.path.exists(tempDir):
+                shutil.rmtree(tempDir)
+    job.openTempDirs = []
 
 
 def main(args=None, stdout=sys.stdout):
